backend/pyproject.toml
[project]
name = "vae-medical-compression"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "pytorch-msssim>=1.0.0",
    "scikit-image>=0.24.0",
    "Pillow>=10.0.0",
    "numpy>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
